moved parsing function to extra file
